cleaning up player state types on server
