+++ server/spotify/spotify-web-api.ts
@@ -90,7 +90,7 @@ export class SpotifyWebApi {
    * @param accessToken
    * @returns
    */
-  async getPlayerState(accessToken: string): Promise<PlayerState> {
+  async getPlayerState(accessToken: string): Promise<Types.PlayerState> {
     const url = SpotifyWebApi.url('/me/player');
     return await axios
       .get(url, {
